moving named.jsx files to ./named/index.js format, and updating require links
