+++ client/my-sites/stats/action-spam/index.jsx
