+++ client/my-sites/stats/info-panel/index.jsx
