+++ client/my-sites/stats/insights/index.jsx
@@ -7,18 +7,18 @@ var React = require( 'react/addons' ),
 /**
  * Internal dependencies
  */
-var StatsNavigation = require( './stats-navigation' ),
+var StatsNavigation = require( '../stats-navigation' ),
 	SidebarNavigation = require( 'my-sites/sidebar-navigation' ),
 	AllTime = require( 'my-sites/stats/all-time/' ),
-	Comments = require( './module-comments' ),
-	Followers = require( './module-followers' ),
+	Comments = require( '../modules/module-comments' ),
+	Followers = require( '../modules/module-followers' ),
 	PostTrends = require( 'my-sites/post-trends' ),
-	SiteOverview = require( './overview' ),
-	SiteOverviewPlaceholder = require( 'my-sites/stats/module-site-overview-placeholder' ),
-	StatsModule = require( './stats-module' ),
-	statsStrings = require( './stats-strings' ),
+	SiteOverview = require( '../overview' ),
+	SiteOverviewPlaceholder = require( 'my-sites/stats/modules/module-site-overview-placeholder' ),
+	StatsModule = require( '../stats-module' ),
+	statsStrings = require( '../stats-strings' ),
 	MostPopular = require( 'my-sites/stats/most-popular' ),
-	PostPerformance = require( './post-performance' ),
+	PostPerformance = require( '../post-performance' ),
 	analytics = require( 'analytics' ),
 	observe = require( 'lib/mixins/data-observe' ),
 	touchDetect = require( 'lib/touch-detect' ),
