+++ client/my-sites/stats/modules/module-chart-tabs.jsx
