+++ client/my-sites/stats/modules/module-countries.jsx
@@ -8,13 +8,13 @@ var React = require( 'react' ),
 /**
  * Internal dependencies
  */
-var toggle = require( './mixin-toggle' ),
-	Geochart = require( './geochart' ),
-	StatsList = require( './stats-list' ),
+var toggle = require( '../mixin-toggle' ),
+	Geochart = require( '../geochart' ),
+	StatsList = require( '../stats-list' ),
 	observe = require( 'lib/mixins/data-observe' ),
 	ErrorPanel = require( './module-error' ),
-	skeleton = require( './mixin-skeleton' ),
-	DownloadCsv = require( './download-csv' ),
+	skeleton = require( '../mixin-skeleton' ),
+	DownloadCsv = require( '../download-csv' ),
 	DatePicker = require( './module-date-picker' ),
 	Card = require( 'components/card' ),
 	Gridicon = require( 'components/gridicon' );
