+++ client/my-sites/stats/modules/module-date-picker.jsx
