+++ client/my-sites/stats/modules/module-followers-page.jsx
@@ -8,12 +8,12 @@ var React = require( 'react' ),
 /**
  * Internal dependencies
  */
-var StatsList = require( './stats-list' ),
+var StatsList = require( '../stats-list' ),
 	SelectDropdown = require( 'components/select-dropdown' ),
-	toggle = require( './mixin-toggle' ),
-	skeleton = require( './mixin-skeleton' ),
+	toggle = require( '../mixin-toggle' ),
+	skeleton = require( '../mixin-skeleton' ),
 	ErrorPanel = require( './module-error' ),
-	Pagination = require( './pagination' ),
+	Pagination = require( '../pagination' ),
 	analytics = require( 'analytics' ),
 	Card = require( 'components/card' ),
 	Gridicon = require( 'components/gridicon' );
