+++ client/my-sites/stats/modules/module-summary-chart.jsx
