+++ client/my-sites/stats/modules/module-tab.jsx
