+++ client/my-sites/stats/modules/module-tabs.jsx
