+++ client/my-sites/stats/modules/module-video-details.jsx
@@ -8,8 +8,8 @@ var React = require( 'react' ),
 /**
  * Internal dependencies
  */
-var StatsList = require( './stats-list' ),
-	skeleton = require( './mixin-skeleton' ),
+var StatsList = require( '../stats-list' ),
+	skeleton = require( '../mixin-skeleton' ),
 	Card = require( 'components/card' );
