+++ client/my-sites/stats/site/index.jsx
@@ -10,15 +10,15 @@ var page = require( 'page' ),
 /**
  * Internal dependencies
  */
-var StatsNavigation = require( './stats-navigation' ),
+var StatsNavigation = require( '../stats-navigation' ),
 	SidebarNavigation = require( 'my-sites/sidebar-navigation' ),
-	DatePicker = require( './module-date-picker' ),
-	Countries = require( './module-countries' ),
-	Comments = require( './module-comments' ),
-	Followers = require( './module-followers' ),
-	ChartTabs = require( './module-chart-tabs' ),
-	StatsModule = require( './stats-module' ),
-	statsStrings = require( './stats-strings' ),
+	DatePicker = require( '../modules/module-date-picker' ),
+	Countries = require( '../modules/module-countries' ),
+	Comments = require( '../modules/module-comments' ),
+	Followers = require( '../modules/module-followers' ),
+	ChartTabs = require( '../modules/module-chart-tabs' ),
+	StatsModule = require( '../stats-module' ),
+	statsStrings = require( '../stats-strings' ),
 	titlecase = require( 'to-title-case' ),
 	analytics = require( 'analytics' ),
 	config = require( 'config' ),
