+++ client/my-sites/stats/stats-overview/index.jsx
@@ -10,10 +10,10 @@ var React = require( 'react' ),
  */
 var observe = require( 'lib/mixins/data-observe' ),
 	SidebarNavigation = require( 'my-sites/sidebar-navigation' ),
-	SiteOverview = require( './overview' ),
-	SiteOverviewPlaceholder = require( './module-site-overview-placeholder' ),
-	DatePicker = require( './module-date-picker' ),
-	StatsNavigation = require( './stats-navigation' ),
+	SiteOverview = require( '../overview' ),
+	SiteOverviewPlaceholder = require( '../modules/module-site-overview-placeholder' ),
+	DatePicker = require( '../modules/module-date-picker' ),
+	StatsNavigation = require( '../stats-navigation' ),
 	analytics = require( 'analytics' ),
 	user = require( 'lib/user' )(),
 	Main = require( 'components/main' ),
